wip abstract method to request service
